3.04 Egli model and unit test

+++ models/sui.hh
@@ -1,6 +1,6 @@
 #ifndef _SUI_HH_
 #define _SUI_HH_
 
-double SUIpathLoss(float f, float TxH, float RxH, float d, int mode);
+double SUIpathLoss(double f, double TxH, double RxH, double d, int mode);
 
 #endif /* _SUI_HH_ */
